+++ rebdev.pavel/T1/isoScale.cpp
@@ -2,7 +2,7 @@
 
 void rebdev::isoScale(Shape ** shapes, const int & numOfShape, const point_t & scalePoint, const double & k)
 {
-  for (int i = 0; i < numOfShape; ++i)
+  for (size_t i = 0; i < numOfShape; ++i)
   {
     point_t centerPoint = shapes[i]->getFrameRect().pos;
     shapes[i]->move(scalePoint);
